This is a new program, version 2 of old and crusty glomtogether.py
Uses Pandas instead of NumPy.  One by-product is that the header is no longer
CSI format.  Also the ISO time stamps drop the double-qoutes.

At the moment you can use the key files for the old glomtogether.py with
this program but down the road we'll move past that as there is a fair amount
of cruft and complication.  JSON / YAML etc.