Variety of small bug fixes.  I should make a broader note on this but the highest resolution is one minute output.  There is some weirdness that just started when libre office spreadsheet files replaced old timey xls files. In the spreadsheet A15=A14+1/24 is getting computed to sometimes end up as a 00:59 instead of 1:00.  Inconsistently but I think I maybe just hadn't noticed till now?  unclear.  Anyway, to fix that every once in a while need to enter: A15=14+1/23.999 in a cell that flips to 00:59.  That is sloppy enough in the spreadsheet to keep things on track.  But at the cost of the seconds data.
Also dropped some of the pieces that use the arrow library.  I need to follow
up on this but I think it was importing the times from the spreadsheet
as local to this timezone rather than UTC and I didn't go step by step to see
all of the spots where this was happening since there were a few.

Anyway, can be revisted later as well.